+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# Version 1.0.0 (wip)
+# Version 1.0.1 (wip)
+
+# Version 1.0.0
 - Ported as plsql fucntion set
 - Ported Spring boot interface
-- New listQueues() metrhod to get the list of all queues
+- New listQueues() method  to get the list of all queues
+- New getMetrics() methods to get metrics details